+++ demo/client/src/pages/piechart.js
@@ -1,0 +1,9 @@
+import {charts} from 'metronic-extension';
+
+// Initialize chart.
+const piechart = new charts.PieChart(document.getElementById('piechart'), {
+  width: 400,
+  ajax: {
+    url: 'json/pie-chart.json'
+  },
+});