+++ types/components/charts/index.d.ts
@@ -1,1 +1,2 @@
 export { default as BarChart } from '~/components/charts/BarChart';
+export { default as PieChart } from '~/components/charts/PieChart';